requirements.txt
openai>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tqdm>=4.65.0
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.1
langchain-core>=0.1.0
fastapi>=0.104.1
uvicorn>=0.24.0
chromadb>=0.4.0
pypdf>=3.17.1
tiktoken>=0.5.2
pydantic>=2.5.1
colorama>=0.4.6 